+++ crates/wdl-lsp/src/server.rs
@@ -255,6 +255,15 @@ impl Server {
         let exceptions = options.exceptions.clone();
         let ignore_name = options.ignore_filename.clone();
         let analyzer_client = client.clone();
+
+        let mut all_rules: Vec<_> = wdl_analysis::rules()
+            .iter()
+            .map(|r| r.id().to_string())
+            .chain(wdl_lint::rules().iter().map(|r| r.id().to_string()))
+            .collect();
+        all_rules.sort_unstable();
+        all_rules.dedup();
+
         // TODO ACF 2025-07-07: add configurability around the fallback behavior; see
         // https://github.com/stjude-rust-labs/wdl/issues/517
         let analyzer_config = AnalysisConfig::default()
@@ -264,7 +273,8 @@ impl Server {
                     .iter()
                     .filter(|r| exceptions.contains(&r.id().into())),
             ))
-            .with_ignore_filename(ignore_name);
+            .with_ignore_filename(ignore_name)
+            .with_all_rules(all_rules);
 
         Self {
             client,
@@ -421,7 +431,11 @@ impl LanguageServer for Server {
                 references_provider: Some(OneOf::Left(true)),
                 completion_provider: Some(CompletionOptions {
                     resolve_provider: Some(false),
-                    trigger_characters: Some(vec![".".to_string(), "[".to_string()]),
+                    trigger_characters: Some(vec![
+                        ".".to_string(),
+                        "[".to_string(),
+                        "#".to_string(),
+                    ]),
                     ..Default::default()
                 }),
                 hover_provider: Some(HoverProviderCapability::Simple(true)),